# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CMMS API — Beit Jala Pharmaceutical Co.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /machines:
    get:
      operationId: getMachines
      tags: [machines]
      summary: List all active machines
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by machine name or machine number
      responses:
        "200":
          description: List of machines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Machine"
    post:
      operationId: createMachine
      tags: [machines]
      summary: Create a new machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineInput"
      responses:
        "201":
          description: Machine created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /machines/{id}:
    get:
      operationId: getMachine
      tags: [machines]
      summary: Get machine by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Machine found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "404":
          description: Machine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMachine
      tags: [machines]
      summary: Update a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MachineUpdate"
      responses:
        "200":
          description: Machine updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "404":
          description: Machine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /machines/{id}/soft-delete:
    patch:
      operationId: softDeleteMachine
      tags: [machines]
      summary: Soft delete a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Machine soft-deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "404":
          description: Machine not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /machines/{id}/equipment-information:
    get:
      operationId: getEquipmentInformation
      tags: [equipment]
      summary: Get equipment information record for a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment information record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentInformation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertEquipmentInformation
      tags: [equipment]
      summary: Create or update equipment information record for a machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInformationInput"
      responses:
        "200":
          description: Equipment information saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentInformation"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Machine:
      type: object
      required: [id, machineNumber, machineName, status, createdAt]
      properties:
        id:
          type: integer
        machineNumber:
          type: string
        machineName:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
        pmFrequencyMonths:
          type: ["integer", "null"]
        pmStartDate:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    MachineInput:
      type: object
      required: [machineNumber, machineName]
      properties:
        machineNumber:
          type: string
          minLength: 1
        machineName:
          type: string
          minLength: 1
        departmentId:
          type: ["integer", "null"]
        location:
          type: string
        status:
          type: string
        pmFrequencyMonths:
          type: ["integer", "null"]
        pmStartDate:
          type: ["string", "null"]
    MachineUpdate:
      type: object
      properties:
        machineNumber:
          type: string
        machineName:
          type: string
        departmentId:
          type: ["integer", "null"]
        location:
          type: string
        status:
          type: string
        pmFrequencyMonths:
          type: ["integer", "null"]
        pmStartDate:
          type: ["string", "null"]
    EquipmentInformation:
      type: object
      required: [id, machineId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        machineId:
          type: integer
        nameOfEquipment:
          type: ["string", "null"]
        modelNumber:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        identificationNumber:
          type: ["string", "null"]
        datePurchased:
          type: ["string", "null"]
        purchasedFromName:
          type: ["string", "null"]
        purchasedFromAddress:
          type: ["string", "null"]
        manufacturingCompanyName:
          type: ["string", "null"]
        manufacturingCompanyAddress:
          type: ["string", "null"]
        dimensionWidthCm:
          type: ["number", "null"]
        dimensionHeightCm:
          type: ["number", "null"]
        dimensionDepthCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        utilitiesPowerSupply:
          type: ["string", "null"]
        utilitiesAir:
          type: ["string", "null"]
        utilitiesWater:
          type: ["string", "null"]
        utilitiesOther:
          type: ["string", "null"]
        others:
          type: ["string", "null"]
        safetyIssues:
          type: ["string", "null"]
        preparedByName:
          type: ["string", "null"]
        preparedByDate:
          type: ["string", "null"]
        approvedByName:
          type: ["string", "null"]
        approvedByDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EquipmentInformationInput:
      type: object
      properties:
        nameOfEquipment:
          type: string
        modelNumber:
          type: string
        serialNumber:
          type: string
        identificationNumber:
          type: string
        datePurchased:
          type: string
        purchasedFromName:
          type: string
        purchasedFromAddress:
          type: string
        manufacturingCompanyName:
          type: string
        manufacturingCompanyAddress:
          type: string
        dimensionWidthCm:
          type: ["number", "null"]
        dimensionHeightCm:
          type: ["number", "null"]
        dimensionDepthCm:
          type: ["number", "null"]
        weightKg:
          type: ["number", "null"]
        utilitiesPowerSupply:
          type: string
        utilitiesAir:
          type: string
        utilitiesWater:
          type: string
        utilitiesOther:
          type: string
        others:
          type: string
        safetyIssues:
          type: string
        preparedByName:
          type: string
        preparedByDate:
          type: string
        approvedByName:
          type: string
        approvedByDate:
          type: string
...