# lib/api-zod/src/generated/types/machine.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CMMS API — Beit Jala Pharmaceutical Co.
 * OpenAPI spec version: 0.1.0
 */

export interface Machine {
  id: number;
  machineNumber: string;
  machineName: string;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  /** @nullable */
  location?: string | null;
  status: string;
  /** @nullable */
  pmFrequencyMonths?: number | null;
  /** @nullable */
  pmStartDate?: string | null;
  /** @nullable */
  deletedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}