# lib/api-zod/src/generated/types/machineUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CMMS API — Beit Jala Pharmaceutical Co.
 * OpenAPI spec version: 0.1.0
 */

export interface MachineUpdate {
  machineNumber?: string;
  machineName?: string;
  /** @nullable */
  departmentId?: number | null;
  location?: string;
  status?: string;
  /** @nullable */
  pmFrequencyMonths?: number | null;
  /** @nullable */
  pmStartDate?: string | null;
}